feat(refactor-find-inactive-workplaces): Re-structure non route files
